ml_service/requirements.txt
fastapi
uvicorn
torch
transformers
pillow
python-multipart
